# openapi.yaml
openapi: 3.0.3
info:
  title: Personal Tracker API
  version: 0.1.0
  description: REST API for the Personal Tracker expense and budget tracking app.
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /expenses:
    post:
      tags: [Expenses]
      summary: Log an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
    get:
      tags: [Expenses]
      summary: List your expenses
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
...
  /expenses/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Expenses]
      summary: Get an expense
      responses:
        '200':
          description: Expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Expenses]
      summary: Update an expense
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '200':
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Expenses]
      summary: Soft-delete an expense
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /recurring-expenses:
    post:
      tags: [Recurring Expenses]
      summary: Create a recurring expense rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseInput'
      responses:
        '201':
          description: Recurring expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpense'
    get:
      tags: [Recurring Expenses]
      summary: List your recurring expense rules
      responses:
        '200':
          description: Recurring expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringExpense'
  /recurring-expenses/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Recurring Expenses]
      summary: Get a recurring expense rule
      responses:
        '200':
          description: Recurring expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpense'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Recurring Expenses]
      summary: Update a recurring expense rule (including pausing via isActive)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseUpdateInput'
      responses:
        '200':
          description: Updated recurring expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpense'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Recurring Expenses]
      summary: Soft-delete a recurring expense rule
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications:
    post:
      tags: [Notifications]
      summary: Create a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationInput'
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
    get:
      tags: [Notifications]
      summary: List your notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
...
  /notifications/{id}/read:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [Notifications]
      summary: Mark one notification as read
      responses:
        '200':
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    delete:
      tags: [Notifications]
      summary: Delete a notification
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /savings-goals:
    post:
      tags: [Savings Goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalInput'
      responses:
        '201':
          description: Savings goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
    get:
      tags: [Savings Goals]
      summary: List your savings goals
      responses:
        '200':
          description: Savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavingsGoal'
  /savings-goals/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Savings Goals]
      summary: Get a savings goal
      responses:
        '200':
          description: Savings goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Savings Goals]
      summary: Update a savings goal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalUpdateInput'
      responses:
        '200':
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Savings Goals]
      summary: Soft-delete a savings goal
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /savings-goals/{id}/contribute:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Savings Goals]
      summary: Add (or, with a negative amount, withdraw) funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  description: Positive to contribute, negative to withdraw
      responses:
        '200':
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
        '400':
          description: Would take the balance below 0
        '404':
          $ref: '#/components/responses/NotFound'
  /households:
    post:
      tags: [Households]
      summary: Create a household (creator becomes owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Household created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
    get:
      tags: [Households]
      summary: List households you belong to
      responses:
        '200':
          description: Households
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Household'
  /households/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Households]
      summary: Get a household (members only)
      responses:
        '200':
          description: Household
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Households]
      summary: Soft-delete a household (owner only)
      responses:
        '204':
          description: Deleted
        '403':
          description: Not the household owner
        '404':
          $ref: '#/components/responses/NotFound'
  /households/{id}/members:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Households]
      summary: Add a member by email (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Updated household
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '403':
          description: Not the household owner
        '404':
          description: Household not found, or no user with that email
  /households/{id}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Households]
      summary: Remove a member (owner, or the member removing themselves)
      responses:
        '204':
          description: Removed
        '400':
          description: The owner cannot be removed this way
        '403':
          description: Not permitted to remove this member
  /households/{id}/budgets:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [Households]
      summary: Create the shared household budget for a month (any member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, month, year]
              properties:
                amount:
                  type: number
                  minimum: 0.01
                month:
                  type: integer
                  minimum: 1
                  maximum: 12
                year:
                  type: integer
      responses:
        '201':
          description: Household budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdBudget'
        '409':
          description: A household budget already exists for this period
    get:
      tags: [Households]
      summary: List the household's budgets
      responses:
        '200':
          description: Household budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdBudget'
  /households/{id}/budgets/summary:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Households]
      summary: Budget vs. actual spend across all household members for a month
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Household budget summary, broken down per member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdBudgetSummary'
  /households/{id}/budgets/{budgetId}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
      - name: budgetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Households]
      summary: Update the household budget's amount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  minimum: 0.01
      responses:
        '200':
          description: Updated household budget
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdBudget'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Households]
      summary: Soft-delete a household budget
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found or not owned by the caller
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
...
    ExpenseInput:
      type: object
      required: [categoryId, amount, date]
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
        date:
          type: string
          format: date-time
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        amount:
          type: string
          description: Decimal amount, serialized as a string
        description:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        recurringExpenseId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    RecurringExpenseInput:
      type: object
      required: [categoryId, amount, frequency, startDate]
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    RecurringExpenseUpdateInput:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        endDate:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
    RecurringExpense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        amount:
          type: string
        description:
          type: string
          nullable: true
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY]
        startDate:
          type: string
          format: date-time
        nextRunDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
    NotificationInput:
      type: object
      required: [type, title, message]
      properties:
        type:
          type: string
          enum: [BUDGET_THRESHOLD, RECURRING_EXPENSE, SYSTEM]
        title:
          type: string
        message:
          type: string
        metadata:
          type: object
          additionalProperties: true
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
          enum: [BUDGET_THRESHOLD, RECURRING_EXPENSE, SYSTEM]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    SavingsGoalInput:
      type: object
      required: [name, targetAmount]
      properties:
        name:
          type: string
        targetAmount:
          type: number
          minimum: 0.01
        targetDate:
          type: string
          format: date-time
    SavingsGoalUpdateInput:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
          minimum: 0.01
        targetDate:
          type: string
          format: date-time
          nullable: true
    SavingsGoal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        targetAmount:
          type: string
        currentAmount:
          type: string
        targetDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
    HouseholdMemberUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
    HouseholdMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [OWNER, MEMBER]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/HouseholdMemberUser'
    Household:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ownerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: '#/components/schemas/HouseholdMember'
    HouseholdBudget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        amount:
          type: string
        month:
          type: integer
        year:
          type: integer
        lastNotifiedThreshold:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
    HouseholdBudgetSummary:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        budgetAmount:
          type: number
          nullable: true
        actualSpent:
          type: number
        remaining:
          type: number
          nullable: true
        percentUsed:
          type: number
          nullable: true
        byMember:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                format: uuid
              name:
                type: string
              spent:
                type: number